chore: move update-details section slice to slices
