+++ src/widgets/BhandaraCheckin/store/actions/resetAppState.ts
@@ -1,7 +1,7 @@
 import { AnyAction, Dispatch } from "@reduxjs/toolkit";
 import { bhandaraCheckinSlice } from "../../BhandaraCheckin/bhandaraCheckinSlice";
 import { mainSectionSlice } from "../../SectionMain/mainSectionSlice";
-import { updateDetailsSectionSlice } from "../../SectionUpdateDetails/updateDetailsSectionSlice";
+import { updateDetailsSectionSlice } from "../slices/updateDetailsSectionSlice";
 
 export const resetAppState = (dispatch: Dispatch<AnyAction>) => {
   dispatch(mainSectionSlice.actions.reset());
