+++ src/widgets/BhandaraCheckin/store/actions/startCheckin/startCheckinAbhyasi.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 import { RootState, ThunkApiConfig } from "../../index";
 import { bhandaraCheckinSlice } from "../../../BhandaraCheckin/bhandaraCheckinSlice";
 import { mainSectionSlice } from "../../../SectionMain/mainSectionSlice";
-import { updateDetailsSectionSlice } from "../../../SectionUpdateDetails/updateDetailsSectionSlice";
+import { updateDetailsSectionSlice } from "../../slices/updateDetailsSectionSlice";
 import { UserSRCM } from "../../../types";
 import {
   AttendanceExistEnumType,
