+++ src/widgets/BhandaraCheckin/store/index.ts
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import { snackbarSlice } from "../../../components/Snackbar/snackbarSlice";
 import { BhandaraCheckinAPIs } from "../types";
 import { mainSectionSlice } from "../SectionMain/mainSectionSlice";
-import { updateDetailsSectionSlice } from "../SectionUpdateDetails/updateDetailsSectionSlice";
+import { updateDetailsSectionSlice } from "./slices/updateDetailsSectionSlice";
 import { bhandaraCheckinSlice } from "../BhandaraCheckin/bhandaraCheckinSlice";
 
 export const rootReducer = {
